+++ CHANGES.md
@@ -7,6 +7,7 @@
 * added filter for setting API URL data
 * added filter for setting API remote install data
 * added filters for setting API language pack data
+* remove `Settings::set_auth_required()`, now set in API plugins
 
 #### 9.9.10 / 2021-02-18
 * fix change to `redirect_on_save()`